+++ scqubits/core/namedslots_array.py
@@ -590,7 +590,7 @@ class NamedSlotsNdarray(np.ndarray, Serializable):
 
         typename = "NamedSlotsNdarray"
         io_attributes = None
-        if self.dtype in [np.float64, np.complex64, np.int_]:
+        if self.dtype in [np.float64, np.float128, np.int_]:
             io_ndarrays: Optional[Dict[str, ndarray]] = {
                 "input_array": self.view(np.ndarray)
             }
